fix: move team vault folder into management dialog
